refactor: mutations
