+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,5 @@
 import { getClient } from "@app/_components/ApolloClientRSC";
-import { CREATE_AUTH } from "@app/lib/graphql/mutations";
+import { auth } from "@app/lib/graphql/mutations";
 import { GET_USER } from "@app/lib/graphql/queries";
 import NextAuth from "next-auth";
 import KakaoProvider from "next-auth/providers/kakao";
@@ -20,7 +20,7 @@ const handler = NextAuth({
   callbacks: {
     async signIn({ user }) {
       await getClient().mutate({
-        mutation: CREATE_AUTH,
+        mutation: auth.CREATE_AUTH,
         variables: { ...user },
       });
       return true;
