+++ src/app/lib/graphql/mutations/budget.ts
@@ -1,0 +1,43 @@
+import { gql } from "apollo-server-micro";
+
+const CREATE_BUDGET = gql`
+  mutation CreateBudget(
+    $title: String
+    $currencyId: String
+    $type: String
+    $tripId: ID
+  ) {
+    createBudget(
+      title: $title
+      type: $type
+      currencyId: $currencyId
+      tripId: $tripId
+    ) {
+      id
+      title
+      type
+      Currency {
+        id
+        name
+        amountUnit
+      }
+      expenses {
+        amount
+      }
+      incomes {
+        amount
+        exchangeRate
+      }
+    }
+  }
+`;
+
+const DELETE_BUDGET = gql`
+  mutation DeleteBudget($id: ID!) {
+    deleteBudget(id: $id) {
+      id
+    }
+  }
+`;
+
+export default { CREATE_BUDGET, DELETE_BUDGET };
