+++ src/app/lib/graphql/mutations/index.ts
@@ -1,0 +1,6 @@
+import budget from "./budget";
+import expense from "./expense";
+import income from "./income";
+import trip from "./trip";
+import auth from "./auth";
+export { budget, expense, income, trip, auth };
