+++ src/app/lib/graphql/mutations/trip.ts
@@ -1,0 +1,41 @@
+import { gql } from "apollo-server-micro";
+
+const CREATE_TRIP = gql`
+  mutation CreateTrip(
+    $title: String
+    $startedAt: Date
+    $endedAt: Date
+    $countryId: String
+    $userId: ID
+  ) {
+    createTrip(
+      title: $title
+      startedAt: $startedAt
+      endedAt: $endedAt
+      countryId: $countryId
+      userId: $userId
+    ) {
+      id
+      title
+      startedAt
+      endedAt
+      countryId
+      Country {
+        id
+        name
+        name_en
+        flag_img
+        currencyId
+      }
+    }
+  }
+`;
+
+const DELETE_TRIP = gql`
+  mutation DeleteTrip($id: ID!) {
+    deleteTrip(id: $id) {
+      id
+    }
+  }
+`;
+export default { CREATE_TRIP, DELETE_TRIP };
